Finished refactoring the fieldSelect saga
