+++ src/redux/sagas/products.saga.js
@@ -1,16 +1,16 @@
 import { put, takeLatest } from 'redux-saga/effects';
 import axios from 'axios';
+import createProductPriceObject from '../../hooks/createProductPriceObject';
 
-// fetchProducts generator to use get from the db
-function* fetchProducts() {
+// fetchProducts generator to get the array of all products data from the DB
+function* fetchProductsArray() {
   try {
 
     const response = yield axios.get('/api/products');
-    // create an object to have keys indicating products with values of the price for that product
 
     // set products used in reducer
     yield put({
-      type: 'SET_PRODUCTS',
+      type: 'SET_PRODUCTS_ARRAY',
       payload: response.data
     });
   } catch (error) {
@@ -18,13 +18,33 @@ function* fetchProducts() {
   }
 }
 
+
+// worker saga to get all products data from DB and create the custom product identifier/price object to be used in the estimate create and lookup views
+function* fetchProductsObject() {
+  try {
+    const response = yield axios.get('/api/products');
+
+    // use the createProductPriceObject function to convert the returned product array into an object
+    const productObject = createProductPriceObject(response.data);
+    
+    // set product object in reducer
+    yield put({
+      type: 'SET_PRODUCTS_OBJECT',
+      payload: productObject
+    });
+  } catch (error) {
+    console.log('error with fetchProducts in field select saga', error);
+  }
+}
+
+
 function* updateProduct(action) {
   try {
     // takes payload to database to update product
     yield axios.put(`/api/products/${action.payload.id}`, action.payload);
     // fetches products
     yield put({
-      type: 'FETCH_PRODUCTS'
+      type: 'FETCH_PRODUCTS_ARRAY'
     });
   } catch (error) {
     console.log('Error in update product saga: ', error);
@@ -37,7 +57,7 @@ function* addProduct(action) {
     // takes company name input payload and posts to database
     yield axios.post('/api/products', action.payload);
     // refresh products with new product post
-    yield put({ type: 'FETCH_PRODUCTS'});
+    yield put({ type: 'FETCH_PRODUCTS_ARRAY'});
   } catch (error) {
     console.log('Error in post product saga:', error);
   }
@@ -45,7 +65,8 @@ function* addProduct(action) {
 
 // companies saga to fetch companies
 function* productsSaga() {
-  yield takeLatest('FETCH_PRODUCTS', fetchProducts);
+  yield takeLatest('FETCH_PRODUCTS_ARRAY', fetchProductsArray);
+  yield takeLatest('FETCH_PRODUCTS_OBJECT', fetchProductsObject)
   yield takeLatest('UPDATE_PRODUCT', updateProduct);
   yield takeLatest('ADD_PRODUCT', addProduct);
 }
